// ts/commander/src/api/schemas/sectionOut.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Corpora API
 * API for managing and processing corpora
 * OpenAPI spec version: 0.1.0
 */
import type { SectionOutIntroduction } from "./sectionOutIntroduction";
import type { SectionOutInstructions } from "./sectionOutInstructions";
import type { SectionOutOrder } from "./sectionOutOrder";

export interface SectionOut {
  title: string;
  introduction?: SectionOutIntroduction;
  instructions?: SectionOutInstructions;
  order?: SectionOutOrder;
  id: string;
  project_id: string;
  created_at: string;
  updated_at: string;
}