// ts/commander/src/api/schemas/sectionOutline.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Corpora API
 * API for managing and processing corpora
 * OpenAPI spec version: 0.1.0
 */
import type { SubsectionOutline } from "./subsectionOutline";

export interface SectionOutline {
  title: string;
  order: number;
  instructions: string;
  subsections: SubsectionOutline[];
}